server/requirements.txt
# Core Frameworks
fastapi
uvicorn

# Database
appwrite

# Authentication and Security
python-jose[cryptography]  # For JWT token encoding/decoding
passlib[bcrypt]            # For secure password hashing
python-dotenv              # For loading .env files

# Data Validation
pydantic

# Billing and Payments
stripe                     # Official Stripe Python SDK

# Utility (Used for accessing the request object for logging IP/User-Agent)
starlette                  # FastAPI dependency, good practice to list explicitly

# AI/NLP Requirements
sumy
ltk
yake
# MISSING DEPENDENCIES REQUIRED BY 'transformers' and others
requests
tqdm